Project1/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = ""
authors = [
    {name = "Akash-kafle",email = "aakashkafle47@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
packages = [
    { include = "project1" }
]
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"